requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
Django==5.0.2
django-nested-admin==4.0.2
django-phonenumber-field==7.3.0
fonttools==4.49.0
html5lib==1.1
idna==3.6
kombu==5.3.5
lxml==5.1.0
oscrypto==1.3.0
phonenumbers==8.13.31
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
pydyf==0.9.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.1.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-monkey-business==1.0.0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.15
zopfli==0.2.3
